+++ src/main/java/christmas/common/exception/ChristmasStateException.java
@@ -1,4 +1,4 @@
-package christmas.exception;
+package christmas.common.exception;
 
 public class ChristmasStateException extends IllegalStateException {
