+++ src/main/java/christmas/common/exception/InvalidMoneyException.java
@@ -1,4 +1,4 @@
-package christmas.exception;
+package christmas.common.exception;
 
 public class InvalidMoneyException extends ChristmasArgumentException {
