+++ src/test/java/christmas/vo/MenuTest.java
@@ -2,6 +2,9 @@ package christmas.vo;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import christmas.common.domain.Menu;
+import christmas.common.domain.MenuType;
+import christmas.common.domain.Money;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
